themed editor use context
