+++ website/package.json
@@ -11,8 +11,8 @@
     "postbuild": "node postprocess"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0-alpha.19",
-    "@docusaurus/preset-classic": "^2.0.0-alpha.19",
+    "@docusaurus/core": "^2.0.0-alpha.43",
+    "@docusaurus/preset-classic": "^2.0.0-alpha.43",
     "classnames": "^2.2.6",
     "loud-rejection": "^2.1.0",
     "monaco-editor": "^0.19.2",
